UPD: changed userID to 64

+++ services/auth/pkg/models.go
@@ -2,11 +2,12 @@ package pkg
 
 import (
 	"fmt"
+
 	"github.com/go-playground/validator/v10"
 )
 
 type User struct {
-	ID       int32  `json:"id"`
+	ID       int64  `json:"id"`
 	Username string `json:"name"     validate:"required,min=3,max=32"`
 	Email    string `json:"email"    validate:"required,email"`
 	Password string `json:"password" validate:"required"`
